+++ ricaun.RevitTest.Tests/TestsRevitIdling.cs
@@ -28,7 +28,6 @@ namespace ricaun.RevitTest.Tests
         {
             Assert.IsNotNull(uiapp);
             Assert.IsTrue(uiapp.IsValidObject);
-            System.Console.WriteLine($"GetHashCode: {this.GetType().Assembly.GetHashCode()}");
         }
 
         [TestCase(1)]
@@ -40,7 +39,6 @@ namespace ricaun.RevitTest.Tests
         {
             Assert.IsNotNull(uiapp);
             Assert.IsTrue(uiapp.IsValidObject);
-            System.Console.WriteLine($"GetHashCode: {this.GetHashCode()}");
         }
     }
 }